+++ R/eng-mod_info_sidebar.R
@@ -72,7 +72,7 @@ eng_mod_info_sidebar_server <- function(id) {
     output$currentPositives <- renderValueBox({
       valueBox(subtitle = "Active cases", icon = icon("info-circle"),
         color = "orange",# fill = TRUE,
-        value = data_to_use()[["totale_attualmente_positivi"]]
+        value = data_to_use()[["totale_positivi"]]
       )
     })