+++ dev/test_mod_ts_ita.R
@@ -6,8 +6,8 @@ golem::document_and_reload()
 ui <- mod_ts_ita_ui("ts", title = "title")
 
 server <- function(input, output ,session){
-    # mod_ts_ita_server("ts", "cum")
-    mod_ts_ita_server("ts", "inc")
+    mod_ts_ita_server("ts", "cum")
+    # mod_ts_ita_server("ts", "inc")
 }
 
 shinyApp(ui, server)