+++ NEWS.md
@@ -1,6 +1,7 @@
+* data update (including renaming)
 * added focus module 20200331 epidemiologia comparativa Veneto-Piemonte
-* removed unuseful button from plotly
-* added istructions to manage plots
+* removed useless button from Plotly
+* added instructions to manage plots
 * colour and label name for levels (#33)
 
 # covid19ita 0.25.2