+++ NEWS.md
@@ -1,3 +1,9 @@
+* data update (including renaming)
+* added focus module 20200331 epidemiologia comparativa Veneto-Piemonte
+* removed useless button from Plotly
+* added instructions to manage plots
+* colour and label name for levels (#33)
+
 # covid19ita 0.25.2
 
 * death and tests labels switched (#34)